requirements.txt
opencv-python
opencv-python-headless
Pillow
pysqlite3
